backend/requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Ma’lumotlar va sozlamalar
pydantic>=2.0
pydantic-settings>=2.0

# Ma’lumotlar bazasi (PostgreSQL, async)
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
alembic>=1.13

# Kesh / navbatlar
redis>=5.0

# HTTP mijoz (tashqi APIlar)
httpx>=0.27

# Autentifikatsiya (keyinroq JWT)
python-jose[cryptography]>=3.3
bcrypt>=4.0,<5.0
python-multipart>=0.0.9
